requirements.txt
fastapi==0.95.0
motor==2.5.1
passlib==1.7.4
python-jose==3.3.0
uvicorn==0.23.1
